Update event.md - the description is hard to understand (#42509)
* Update event.md

The description is hard to understand. It uses circular references to the word event and refers to an undefined term, namely `pub;lisher class".

* Update docs/csharp/language-reference/keywords/event.md

Just suggestions: An ***event*** is a member that enables notifications to be triggered. Clients can attach executable code for events by supplying ***event handlers***. The `event` keyword declares an ***event***. The event is of a delegate type. While triggering an event all supplied event handlers are invoked. Event handlers are delegate instances added to the event. Event users can add or remove their event handlers to/from an event.

Co-authored-by: Bill Wagner <wiwagn@microsoft.com>

---------

Co-authored-by: Bill Wagner <wiwagn@microsoft.com>