Rework introduction on method parameters (#43691)
* Rework parameter passing

Rework the examples and discussions on pass by value and pass by reference.  Remove the (more complicated) discussion of `= ref` and how that impacts modifying parameter values. That's coming in the next commit to the assignment article.

* add ref reassignment

Point out that if you `ref =` assign a parameter to new storage, any subsequent value changes aren't visible outside the method.

* fix build warnings

* Update docs/csharp/language-reference/keywords/snippets/PassParameters.cs