Add more clarity on member lookup
Found while updating the feature spec for the feature. This article didn't clearly state that file local types are preferred over other types with the same name in the file where they are declared.

Also, the `file` modifier doesn't actually modify the *scope*, rather, it modifies the *visibility*. Update the language to match the speclet definition.

Finally, quick edit pass.