Document built-in System.Net activities (#43872)
.NET 9 introduced additions to `System.Net` distributed tracing which define a contract following OTel recommendations:
- https://github.com/dotnet/runtime/pull/104251 adjusted the HTTP client request activity so it natively complies with OTel recommendations. This included adding standard tags.
- https://github.com/dotnet/runtime/pull/103922 introduced activities breaking down the HTTP connection setup.

This PR documents the activities together with their attributes in a manner similar to the metrics docs added by https://github.com/dotnet/docs/pull/37213. The matching semconv PR is https://github.com/open-telemetry/semantic-conventions/pull/1192. Conceptual docs are being introduced in #42830.

Moreover, the PR also adjusts certain parts of the metrics doc in order to make it more accurate, synchronize it with the activities doc and incorporate the changes from #42810.
---------

Co-authored-by: Genevieve Warren <24882762+gewarren@users.noreply.github.com>
Co-authored-by: Miha Zupan <mihazupan.zupan1@gmail.com>