Version updates
Fixes #33554:  Add a link to selecting the UWP version, because UWP uses a different project format.
Fixes #36919: Add that the directory.props.build method should be used only with C# projects, not folders that include both C# and VB projects.
Fixes #41296:  Move the paragraph that explains why the version dropdown is greyed out to the top of the article.