Update SettingsOptions.cs
Changed the Regular Expression to allow an exclamation mark which is used in the sample code.  The existing Regular Expression allowed a dash which was not used in the sample.  I replaced the dash with the exclamation mark. The sample code will now run as is, instead of throwing an exception.  I have tested the updated Regular Expression, and it is working as expected.