Improve dump FAQ
Fixes https://github.com/dotnet/diagnostics/issues/4375

- The previous guidance for diagnosing dump permission issues did not mention the interaction between setcap and ptrace.
I added that information as well as a more complete description of the ptrace access checking.

- I replaced faq-dumps.yml with dump-faq.md. The yml file included markdown style links but that content was not properly
rendered in the html page. Maybe there was a way to make that work as-is but it seemed simpler to use markdown similar
to all the other pages in this part of the docs. I redirected links to the new page but left the old .yml page for now in case
there were external direct links to it.

- There was no link to the FAQ in the TOC so I added one.

- I made some minor edits to the dump overview page that IMO improved clarity and flow of the text.