Freshen string comparison code (#44736)
* Move and update all code

* Fixes #28006

Add an example that checks equality of instance delegates to point out that the receivers must be the same for those methods.

* Equate `==` with the `is` pattern matching

To fix #42548, create links between the equality operator and the pattern matching `is` expression, in the case where the right operand is a constant.

In addition, add notes in the "How to compare strings" article to bring these together.

* Add more overloads

Fixes #20123  Add additional overloads as necessary, and explain the settings that can be applied among the different overloads.

* final edit pass

* Apply suggestions from code review

Co-authored-by: Cam Soper <cam.soper@microsoft.com>

* respond to feedback.

---------

Co-authored-by: Cam Soper <cam.soper@microsoft.com>