Update notes on IHttpClientFactory and cookies (#45658)
There seems to be a confusion around the reasons why we advise against HttpClientFactory when the app needs to use cookies. https://github.com/dotnet/dotnet-api-docs/issues/10770 and https://github.com/dotnet/dotnet-api-docs/issues/1535 imply that users often think the problem might be related to `CookieContainer` thread safety which is not the case.

This PR updates the notes, describing the actual reasons: cookie leaking and unexpected recycling of cookies.