Update projection-operations (Zip section).md
there is a typo issue , right below the last code snippet in the Zip section, the word 'numbers' should be replaced with 
word 'number'. The sentence is :

With the preceding Zip overload, the specified function is applied to the corresponding elements numbers and letter, producing a sequence of the string results.