Language reference and fundamentals for file based programs (#46900)
* publish spec and add preprocessor reference

First commit: Publish the C# speclet for ignored preprocessor directives, and make the updates for the ignored directives used for file based programs.

* Add file-based programs to get started

* Fix host location

And, set +x permissions on hello-world.cs

* Include file based programs in the structure

Include a description and small example of a file based program in the overview of the general structure of a C# program.

* Add file based programs to the command line

Add a brief description of file based programs in the command line section.

* Updates on top level statements

Final edits.

* Apply suggestions from code review

Co-authored-by: Genevieve Warren <24882762+gewarren@users.noreply.github.com>
Co-authored-by: David Pine <david.pine@microsoft.com>

* apply feedback.

---------

Co-authored-by: Genevieve Warren <24882762+gewarren@users.noreply.github.com>
Co-authored-by: David Pine <david.pine@microsoft.com>