define freshness tiers for F#
Define freshness tiers for different folders in the F# guide.

I followed the general strategy I used for C#:  Areas that are likely to need updates for each release based on new features are 1 year. Most reference is 3 years: Someone should give it a look, even if it's not likely that new features impact it.  Some of the reference for stable features is 5 years. Compiler diagnostics are 10 years.