Fix increment and decrement static operator (#48627)
The increment and decrement static operator currently are no-op. This fix ensure they actually do increment and decrement the value of the numerator in the fraction.

Change the operator from an increment to a non mutating operand.numerator + 1 because otherwise the call to a reference version of Fraction would result in `Console.WriteLine(a++);` to print a after its mutation instead of the before value.

Fixes #48626