Add tip for interpolated string parsing
Fixes #43758

Add an include file with a tip that explains how Visual Studio and the C# Dev Kit interpret raw string literals and provides syntax coloring or potential fixes when the literal doesn't represent valid syntax for the chosen format.

Add that include file in articles that provide information on raw string literals.

Perform an edit pass and freshness pass on impacted articles.