Add note on synthesized ctors
Add a note on the synthesized parameterless constructor.

Fixes #49412

I didn't directly describe the situation in the original issue. Instead, I explained the `record` behavior to emphasize the synthesized constructors. From that, the normal rules follow which could make it impossible to instantiate an instance of a type because no accessible constructors are available.